Changing pages to ssr part 7: lunch order [nodeploy]
